Se cambia los tipos de usuario y se movio el hitorial de pedidos a usaurio

+++ src/main/java/com/licoreria/app/modelo/pedidos/DetalleElementoReceta.java
@@ -4,5 +4,5 @@ import com.licoreria.app.modelo.productos.ElementoReceta;
 
 public class DetalleElementoReceta {
     private ElementoReceta elementoBase;
-    private int cantidadEspecifica;
+    private double cantidadEspecifica;
 }
